+++ tests/test_api.py
@@ -110,13 +110,15 @@ async def test_base_repo_update_updates_row_in_db(repo: TransactionRepo) -> None
     assert updated_trans.date == trans.date
 
 
-async def test_base_repo_first_return_first_matching_row(repo: TransactionRepo, transactions: List[Transaction]) -> None:
+async def test_base_repo_first_return_first_matching_row(repo: TransactionRepo,
+                                                         transactions: List[Transaction]) -> None:
     trans = await repo.first(transactions_table.c.price == 100)
 
     assert trans.id == transactions[0].id
 
 
-async def test_base_repo_get_all_return_all_rows_filtered_and_sorted(repo: TransactionRepo, transactions: List[Transaction]) -> None:
+async def test_base_repo_get_all_return_all_rows_filtered_and_sorted(repo: TransactionRepo,
+                                                                     transactions: List[Transaction]) -> None:
     db_transactions = await repo.get_all(
         filters=[transactions_table.c.price == 100],
         orders=[transactions_table.c.date]
@@ -135,3 +137,28 @@ async def test_transaction_repo_custom_method_works(repo: TransactionRepo, trans
     sum_ = await repo.sum()
 
     assert sum_ == sum(map(operator.attrgetter("price"), transactions))
+
+
+async def test_base_repo_get_by_id_returns_row_with_id(repo: TransactionRepo, transactions: List[Transaction]) -> None:
+    db_trans = await repo.get_by_id(transactions[0].id)
+    assert db_trans == transactions[0]
+
+
+async def test_base_repo_get_or_create_creates_entity_if_no_entities(repo: TransactionRepo) -> None:
+    price = 400
+    trans, created = await repo.get_or_create(defaults={"price": price})
+    assert created
+    assert trans.price == price
+
+
+async def test_base_repo_get_or_create_returns_entity_if_match(
+        repo: TransactionRepo,
+        transactions: List[Transaction]
+) -> None:
+    price = 400
+    trans, created = await repo.get_or_create(
+        filters=[transactions_table.c.id == transactions[0].id],
+        defaults={"price": price}
+    )
+    assert not created
+    assert trans == transactions[0]