+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 0.2.0 - 2019-07-18
+## Unreleased
+
 ### Added
+
+- repka.api.BaseRepository.get_by_id - get entity by id or None
+- repka.api.BaseRepository.get_or_create - get first entity by filters or create new with defaults
 - repka.api.BaseRepository.serializer property for defining serialization behaviour
 - repka.api.BaseRepository.deserializer property for defining deserialization behaviour
-
 ### Removed
 - repka.api.BaseRepository.entity_type property. The same functionality can be provided by deserializer property